requirements.txt
python-telegram-bot[job-queue]==21.10
openai==1.82.0
python-dotenv==1.1.0
aiosqlite==0.21.0
